Emit is working for EASM

+++ EASM/src/instructions.c
@@ -2,8 +2,8 @@
 
 static char * inst_names[] = { "add", "addi", "hcf", "jmp", "lb", "lbi", "li", "lw", "lwi", "mod", "modi", "mov", "pop", "push", "pushi", "sb", "sbi", "sw", "swi", "sub", "subi" };
 
-uint8_t getinst (char * s) {
-    for (uint8_t i = 0; i < 21; i++) {
+int getinst (char * s) {
+    for (int i = 0; i < 21; i++) {
         if (strcmp (s, inst_names [i]) == 0) {
             return i;
         }
